+++ OMICALL-Flutter-SDK-Sample/lib/main.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 
 import 'package:calling/local_storage/local_storage.dart';
 import 'package:calling/screens/home/home_screen.dart';
-import 'package:calling/screens/login/login_apikey_screen.dart';
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
@@ -86,19 +85,10 @@ Future<void> updateToken({
   if (Platform.isIOS) {
     apnToken = await FirebaseMessaging.instance.getAPNSToken();
   }
-  String id = "";
-  DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
-  if (Platform.isAndroid) {
-    id = (await deviceInfo.androidInfo).id;
-  } else {
-    id = (await deviceInfo.iosInfo).identifierForVendor ?? "";
-  }
   if (showLoading) {
     EasyLoading.show();
   }
   await OmicallClient.instance.updateToken(
-    id,
-    Platform.isAndroid ? "omicall.concung.dev" : "vn.vihat.omikit",
     fcmToken: token,
     apnsToken: apnToken,
   );