+++ OMICALL-Flutter-SDK-Sample/lib/screens/login/login_user_password_screen.dart
@@ -25,9 +25,9 @@ class _LoginScreenState extends State<LoginUserPasswordScreen> {
   late final TextEditingController _passwordController = TextEditingController()
     ..text = Platform.isIOS ? 'VlAkzpm2Fn' : 'vWmFFBZwss';
   late final TextEditingController _serviceUrlController =
-      TextEditingController()..text = '';
+      TextEditingController()..text = 'thaonguyennguyen1197';
   late final TextEditingController _hostUrlController = TextEditingController()
-    ..text = '';
+    ..text = '171.244.138.14';
 
   bool _supportVideoCall = true;
   TextStyle basicStyle = const TextStyle(